+++ src/wifi/model/eht/advanced-emlsr-manager.cc
@@ -17,6 +17,7 @@
 #include "ns3/wifi-phy.h"
 
 #include <algorithm>
+#include <sstream>
 
 namespace ns3
 {
@@ -141,7 +142,40 @@ AdvancedEmlsrManager::GetTypeId()
                           "expires, the main PHY is switched back to the preferred link.",
                           TimeValue(MilliSeconds(5)),
                           MakeTimeAccessor(&AdvancedEmlsrManager::m_switchMainPhyBackDelay),
-                          MakeTimeChecker());
+                          MakeTimeChecker())
+            .AddAttribute("KeepMainPhyAfterDlTxop",
+                          "In case aux PHYs are not TX capable and do not switch link, after the "
+                          "end of a DL TXOP carried out on an aux PHY link, the main PHY stays on "
+                          "that link for a switch main PHY back delay, if this attribute is true, "
+                          "or it returns to the preferred link, otherwise.",
+                          BooleanValue(false),
+                          MakeBooleanAccessor(&AdvancedEmlsrManager::m_keepMainPhyAfterDlTxop),
+                          MakeBooleanChecker())
+            .AddAttribute("CheckAccessOnMainPhyLink",
+                          "In case aux PHYs are not TX capable and an Access Category, say it AC "
+                          "X, is about to gain channel access on an aux PHY link, determine "
+                          "whether the time the ACs with priority higher than or equal to AC X and "
+                          "with frames to send on the main PHY link are expected to gain access on "
+                          "the main PHY link should be taken into account when taking the decision "
+                          "to switch the main PHY to the aux PHY link.",
+                          BooleanValue(true),
+                          MakeBooleanAccessor(&AdvancedEmlsrManager::m_checkAccessOnMainPhyLink),
+                          MakeBooleanChecker())
+            .AddAttribute(
+                "MinAcToSkipCheckAccess",
+                "If the CheckAccessOnMainPhyLink attribute is set to false, indicate the "
+                "minimum priority AC for which it is allowed to skip the check related "
+                "to the expected channel access time on the main PHY link.",
+                EnumValue(AcIndex::AC_BK),
+                MakeEnumAccessor<AcIndex>(&AdvancedEmlsrManager::m_minAcToSkipCheckAccess),
+                MakeEnumChecker(AcIndex::AC_BE,
+                                "AC_BE",
+                                AcIndex::AC_VI,
+                                "AC_VI",
+                                AcIndex::AC_VO,
+                                "AC_VO",
+                                AcIndex::AC_BK,
+                                "AC_BK"));
     return tid;
 }
 
@@ -166,10 +200,7 @@ AdvancedEmlsrManager::DoDispose()
             "PhyRxMacHeaderEnd",
             MakeCallback(&AdvancedEmlsrManager::ReceivedMacHdr, this).Bind(phy));
     }
-    if (!GetAuxPhyTxCapable())
-    {
-        GetStaMac()->GetDevice()->GetPhy(GetMainPhyId())->UnregisterListener(m_phyListener);
-    }
+    UnregisterListener();
     m_phyListener.reset();
     DefaultEmlsrManager::DoDispose();
 }
@@ -211,10 +242,30 @@ AdvancedEmlsrManager::DoSetWifiMac(Ptr<StaWifiMac> mac)
             "PhyRxMacHeaderEnd",
             MakeCallback(&AdvancedEmlsrManager::ReceivedMacHdr, this).Bind(phy));
     }
-    if (!GetAuxPhyTxCapable())
+}
+
+void
+AdvancedEmlsrManager::RegisterListener(Ptr<WifiPhy> phy)
+{
+    NS_LOG_FUNCTION(this << phy->GetPhyId());
+
+    NS_ASSERT_MSG(!m_auxPhyWithListener,
+                  "PHY listener is still connected to PHY " << +m_auxPhyWithListener->GetPhyId());
+    phy->RegisterListener(m_phyListener);
+    m_auxPhyWithListener = phy;
+}
+
+void
+AdvancedEmlsrManager::UnregisterListener()
+{
+    if (!m_auxPhyWithListener)
     {
-        mac->GetDevice()->GetPhy(GetMainPhyId())->RegisterListener(m_phyListener);
+        return; // do nothing
     }
+
+    NS_LOG_FUNCTION(this << m_auxPhyWithListener->GetPhyId());
+    m_auxPhyWithListener->UnregisterListener(m_phyListener);
+    m_auxPhyWithListener = nullptr;
 }
 
 std::pair<bool, Time>
@@ -271,7 +322,7 @@ AdvancedEmlsrManager::ReceivedMacHdr(Ptr<WifiPhy> phy,
                                      Time psduDuration)
 {
     auto linkId = GetStaMac()->GetLinkForPhy(phy);
-    if (!linkId.has_value())
+    if (!linkId.has_value() || !m_useNotifiedMacHdr)
     {
         return;
     }
@@ -279,8 +330,8 @@ AdvancedEmlsrManager::ReceivedMacHdr(Ptr<WifiPhy> phy,
 
     auto& ongoingTxopEnd = GetEhtFem(*linkId)->GetOngoingTxopEndEvent();
 
-    if (m_useNotifiedMacHdr && ongoingTxopEnd.IsPending() &&
-        macHdr.GetAddr1() != GetEhtFem(*linkId)->GetAddress() && !macHdr.GetAddr1().IsBroadcast() &&
+    if (ongoingTxopEnd.IsPending() && macHdr.GetAddr1() != GetEhtFem(*linkId)->GetAddress() &&
+        !macHdr.GetAddr1().IsBroadcast() &&
         !(macHdr.IsCts() && macHdr.GetAddr1() == GetEhtFem(*linkId)->GetBssid() /* CTS-to-self */))
     {
         // the EMLSR client is no longer involved in the TXOP and switching to listening mode
@@ -292,7 +343,7 @@ AdvancedEmlsrManager::ReceivedMacHdr(Ptr<WifiPhy> phy,
         // another callback (FEM::ReceivedMacHdr) from the PhyRxMacHeaderEnd trace source of
         // the main PHY, thus invalidating the list of callbacks on which the for loop iterates.
         // Hence, schedule the call to NotifyTxopEnd to execute it outside such for loop.
-        Simulator::ScheduleNow(&AdvancedEmlsrManager::NotifyTxopEnd, this, *linkId, false, false);
+        Simulator::ScheduleNow(&AdvancedEmlsrManager::NotifyTxopEnd, this, *linkId, nullptr);
     }
 
     // if the MAC header has been received on the link on which the main PHY is operating (or on
@@ -301,22 +352,25 @@ AdvancedEmlsrManager::ReceivedMacHdr(Ptr<WifiPhy> phy,
     // expires (plus a channel switch delay), try to switch the main PHY back to the preferred link
     const auto mainPhyInvolved =
         (phy->GetPhyId() == GetMainPhyId()) ||
-        (m_mainPhySwitchInfo.end >= Simulator::Now() && m_mainPhySwitchInfo.to == *linkId);
+        (m_mainPhySwitchInfo.disconnected && m_mainPhySwitchInfo.to == *linkId);
+    const auto delay =
+        Simulator::GetDelayLeft(m_switchMainPhyBackEvent) + phy->GetChannelSwitchDelay();
 
-    if (m_switchMainPhyBackEvent.IsPending() && mainPhyInvolved &&
-        !GetExpectedAccessWithinDelay(*linkId,
-                                      Simulator::GetDelayLeft(m_switchMainPhyBackEvent) +
-                                          phy->GetChannelSwitchDelay()))
-    {
-        m_switchMainPhyBackEvent.Cancel();
-        SwitchMainPhyBackDelayExpired(*linkId);
-    }
+    Simulator::ScheduleNow([=, this]() {
+        if (WifiExpectedAccessReason reason;
+            m_switchMainPhyBackEvent.IsPending() && mainPhyInvolved &&
+            (reason = GetStaMac()->GetChannelAccessManager(*linkId)->GetExpectedAccessWithin(
+                 delay)) != WifiExpectedAccessReason::ACCESS_EXPECTED)
+        {
+            SwitchMainPhyBackDelayExpired(*linkId, reason);
+        }
+    });
 }
 
 void
-AdvancedEmlsrManager::DoNotifyTxopEnd(uint8_t linkId)
+AdvancedEmlsrManager::DoNotifyTxopEnd(uint8_t linkId, Ptr<QosTxop> edca)
 {
-    NS_LOG_FUNCTION(this << linkId);
+    NS_LOG_FUNCTION(this << linkId << edca);
 
     auto mainPhy = GetStaMac()->GetDevice()->GetPhy(m_mainPhyId);
 
@@ -336,6 +390,30 @@ AdvancedEmlsrManager::DoNotifyTxopEnd(uint8_t linkId)
     // - SwitchAuxPhy is true, the main PHY is switching and switching can be interrupted
     // or
     // - SwitchAuxPhy is false and there is an aux PHY to reconnect
+
+    if (!m_auxPhyTxCapable && !m_switchAuxPhy && !edca && m_keepMainPhyAfterDlTxop)
+    {
+        // DL TXOP ended, check if the main PHY must be kept on this link to try to gain an UL TXOP
+        NS_ASSERT_MSG(!m_switchMainPhyBackEvent.IsPending(),
+                      "Switch main PHY back timer should not be running at the end of a DL TXOP");
+        NS_ASSERT_MSG(!mainPhy->IsStateSwitching(),
+                      "Main PHY should not be switching at the end of a DL TXOP");
+
+        if (GetStaMac()->GetChannelAccessManager(linkId)->GetExpectedAccessWithin(
+                m_switchMainPhyBackDelay) == WifiExpectedAccessReason::ACCESS_EXPECTED)
+        {
+            NS_LOG_DEBUG("Keep main PHY on link " << +linkId << " to try to gain an UL TXOP");
+            m_switchMainPhyBackEvent =
+                Simulator::Schedule(m_switchMainPhyBackDelay,
+                                    &AdvancedEmlsrManager::SwitchMainPhyBackDelayExpired,
+                                    this,
+                                    linkId,
+                                    std::nullopt);
+            // start checking PHY activity on the link the main PHY is operating
+            RegisterListener(GetStaMac()->GetWifiPhy(linkId));
+            return;
+        }
+    }
 
     std::shared_ptr<EmlsrMainPhySwitchTrace> traceInfo;
 
@@ -361,7 +439,7 @@ AdvancedEmlsrManager::DoNotifyTxopEnd(uint8_t linkId)
     if (m_switchAuxPhy || !mainPhy->IsStateSwitching() || m_interruptSwitching)
     {
         NS_ASSERT_MSG(
-            !m_switchAuxPhy || m_mainPhySwitchInfo.end >= Simulator::Now(),
+            !m_switchAuxPhy || m_mainPhySwitchInfo.disconnected,
             "Aux PHY next link ID should have a value when interrupting a main PHY switch");
         uint8_t nextLinkId = m_switchAuxPhy ? m_mainPhySwitchInfo.from : GetMainPhyId();
         SwitchMainPhy(nextLinkId, false, REQUEST_ACCESS, std::move(*traceInfo));
@@ -452,29 +530,32 @@ AdvancedEmlsrManager::CheckNavAndCcaLastPifs(Ptr<WifiPhy> phy, uint8_t linkId, P
     auto txopNotStarted = [=, this]() {
         // check when access may be granted to determine whether to switch the main PHY back
         // to the preferred link (if aux PHYs do not switch link)
-        if (!m_switchAuxPhy &&
-            !GetExpectedAccessWithinDelay(linkId,
-                                          Simulator::GetDelayLeft(m_switchMainPhyBackEvent) +
-                                              phy->GetChannelSwitchDelay()))
+        const auto mainPhy = GetStaMac()->GetDevice()->GetPhy(GetMainPhyId());
+        const auto delay =
+            Simulator::GetDelayLeft(m_switchMainPhyBackEvent) + mainPhy->GetChannelSwitchDelay();
+
+        if (WifiExpectedAccessReason reason;
+            !m_switchAuxPhy && m_switchMainPhyBackEvent.IsPending() &&
+            (reason = GetStaMac()->GetChannelAccessManager(linkId)->GetExpectedAccessWithin(
+                 delay)) != WifiExpectedAccessReason::ACCESS_EXPECTED)
         {
             NS_LOG_DEBUG("No AC is expected to get backoff soon, switch main PHY back");
-            if (auto mainPhy = GetStaMac()->GetDevice()->GetPhy(GetMainPhyId());
-                !mainPhy->IsStateSwitching())
-            {
-                m_switchMainPhyBackEvent.Cancel();
-                SwitchMainPhyBackDelayExpired(linkId);
-            }
-            return;
+            SwitchMainPhyBackDelayExpired(linkId, reason);
         }
 
-        // medium busy, restart channel access
+        // restart channel access
         edca->NotifyChannelReleased(linkId); // to set access to NOT_REQUESTED
         edca->StartAccessAfterEvent(linkId,
                                     Txop::DIDNT_HAVE_FRAMES_TO_TRANSMIT,
                                     Txop::CHECK_MEDIUM_BUSY);
     };
 
-    if (!isBusy && width > MHz_u{0})
+    if (!m_switchAuxPhy && !m_switchMainPhyBackEvent.IsPending())
+    {
+        NS_LOG_DEBUG("Main PHY switched back (or scheduled to switch back) before PIFS check");
+        txopNotStarted();
+    }
+    else if (!isBusy && width > MHz_u{0})
     {
         // medium idle, start TXOP
         width = std::min(width, GetChannelForMainPhy(linkId).GetTotalWidth());
@@ -501,15 +582,23 @@ AdvancedEmlsrManager::CheckNavAndCcaLastPifs(Ptr<WifiPhy> phy, uint8_t linkId, P
 }
 
 void
-AdvancedEmlsrManager::SwitchMainPhyBackDelayExpired(uint8_t linkId)
+AdvancedEmlsrManager::SwitchMainPhyBackDelayExpired(
+    uint8_t linkId,
+    std::optional<WifiExpectedAccessReason> stopReason)
 {
-    NS_LOG_FUNCTION(this << linkId);
-
-    if (m_switchAuxPhy)
+    if (g_log.IsEnabled(ns3::LOG_FUNCTION))
     {
-        return; // nothing to do
+        std::stringstream ss;
+        if (stopReason.has_value())
+        {
+            ss << stopReason.value();
+        }
+        NS_LOG_FUNCTION(this << linkId << ss.str());
     }
 
+    m_switchMainPhyBackEvent.Cancel();
+
+    NS_ASSERT_MSG(!m_switchAuxPhy, "Don't expect this to be called when aux PHYs switch link");
     Time extension{0};
 
     // check if the timer must be restarted because a frame is being received on any link
@@ -540,7 +629,8 @@ AdvancedEmlsrManager::SwitchMainPhyBackDelayExpired(uint8_t linkId)
             // this is the link on which the main PHY is operating. If an AC with traffic is
             // expected to get channel access soon (within a channel switch delay), restart
             // the timer to have the main PHY stay a bit longer on this link
-            if (GetExpectedAccessWithinDelay(linkId, phy->GetChannelSwitchDelay()))
+            if (GetStaMac()->GetChannelAccessManager(linkId)->GetExpectedAccessWithin(
+                    phy->GetChannelSwitchDelay()) == WifiExpectedAccessReason::ACCESS_EXPECTED)
             {
                 extension = Max(extension, phy->GetChannelSwitchDelay());
             }
@@ -554,12 +644,51 @@ AdvancedEmlsrManager::SwitchMainPhyBackDelayExpired(uint8_t linkId)
             Simulator::Schedule(extension,
                                 &AdvancedEmlsrManager::SwitchMainPhyBackDelayExpired,
                                 this,
-                                linkId);
+                                linkId,
+                                stopReason);
         return;
     }
 
-    // no need to wait further, switch the main PHY back to the preferred link
-    SwitchMainPhyBackToPreferredLink(linkId, EmlsrSwitchMainPhyBackTrace(false));
+    // no need to wait further, switch the main PHY back to the preferred link and unregister
+    // the PHY listener from the aux PHY
+    const auto elapsed = Simulator::Now() - m_mainPhySwitchInfo.start;
+    const auto isSwitching = GetStaMac()->GetDevice()->GetPhy(GetMainPhyId())->IsStateSwitching();
+    SwitchMainPhyBackToPreferredLink(linkId,
+                                     EmlsrSwitchMainPhyBackTrace(elapsed, stopReason, isSwitching));
+    // if scheduled, invoke CheckNavAndCcaLastPifs(), which will just restart channel access
+    if (m_ccaLastPifs.IsPending())
+    {
+        m_ccaLastPifs.PeekEventImpl()->Invoke();
+        m_ccaLastPifs.Cancel();
+    }
+    UnregisterListener();
+}
+
+void
+AdvancedEmlsrManager::SwitchMainPhyBackToPreferredLink(uint8_t linkId,
+                                                       EmlsrMainPhySwitchTrace&& traceInfo)
+{
+    if (!m_interruptSwitching)
+    {
+        DefaultEmlsrManager::SwitchMainPhyBackToPreferredLink(
+            linkId,
+            std::forward<EmlsrMainPhySwitchTrace>(traceInfo));
+        return;
+    }
+
+    NS_LOG_FUNCTION(this << linkId << traceInfo.GetName());
+
+    NS_ABORT_MSG_IF(m_switchAuxPhy, "This method can only be called when SwitchAuxPhy is false");
+
+    if (!m_auxPhyToReconnect)
+    {
+        return;
+    }
+
+    SwitchMainPhy(GetMainPhyId(),
+                  false,
+                  REQUEST_ACCESS,
+                  std::forward<EmlsrMainPhySwitchTrace>(traceInfo));
 }
 
 void
@@ -578,25 +707,27 @@ AdvancedEmlsrManager::InterruptSwitchMainPhyBackTimerIfNeeded()
 
     if (!linkId.has_value())
     {
-        NS_LOG_DEBUG("Main PHY is not operating on any link");
-        return;
+        NS_ASSERT(m_mainPhySwitchInfo.disconnected);
+        linkId = m_mainPhySwitchInfo.to;
+        NS_LOG_DEBUG("Main PHY is switching to link " << +linkId.value());
     }
 
     const auto delay =
         Simulator::GetDelayLeft(m_switchMainPhyBackEvent) + mainPhy->GetChannelSwitchDelay();
-    if (!GetExpectedAccessWithinDelay(*linkId, delay))
+    if (auto reason = GetStaMac()->GetChannelAccessManager(*linkId)->GetExpectedAccessWithin(delay);
+        reason != WifiExpectedAccessReason::ACCESS_EXPECTED)
     {
-        m_switchMainPhyBackEvent.Cancel();
-        SwitchMainPhyBackDelayExpired(*linkId);
+        SwitchMainPhyBackDelayExpired(*linkId, reason);
     }
 }
 
 void
-AdvancedEmlsrManager::DoNotifyIcfReceived(uint8_t linkId)
+AdvancedEmlsrManager::DoNotifyDlTxopStart(uint8_t linkId)
 {
     NS_LOG_FUNCTION(this << linkId);
     m_switchMainPhyBackEvent.Cancel();
     m_ccaLastPifs.Cancel();
+    UnregisterListener();
 }
 
 void
@@ -605,6 +736,7 @@ AdvancedEmlsrManager::DoNotifyUlTxopStart(uint8_t linkId)
     NS_LOG_FUNCTION(this << linkId);
     m_switchMainPhyBackEvent.Cancel();
     m_ccaLastPifs.Cancel();
+    UnregisterListener();
 }
 
 bool
@@ -614,12 +746,14 @@ AdvancedEmlsrManager::RequestMainPhyToSwitch(uint8_t linkId, AcIndex aci, const 
 
     // the aux PHY is not TX capable; check if main PHY has to switch to the aux PHY's link
     auto mainPhy = GetStaMac()->GetDevice()->GetPhy(m_mainPhyId);
-    const auto mainPhyLinkId = GetStaMac()->GetLinkForPhy(mainPhy);
 
-    // if main PHY is not operating on a link, it is switching, hence do not request another switch
-    if (!mainPhyLinkId.has_value())
+    // if main PHY is not operating on a link and is trying to start a (DL or UL) TXOP, then do
+    // not request another switch
+    if (m_mainPhySwitchInfo.disconnected &&
+        (!m_interruptSwitching || m_mainPhySwitchInfo.reason == "DlTxopIcfReceivedByAuxPhy" ||
+         m_mainPhySwitchInfo.reason == "UlTxopAuxPhyNotTxCapable"))
     {
-        NS_LOG_DEBUG("Main PHY is not operating on any link");
+        NS_LOG_DEBUG("Main PHY is not operating on any link and cannot switch to another link");
         return false;
     }
 
@@ -655,45 +789,74 @@ AdvancedEmlsrManager::RequestMainPhyToSwitch(uint8_t linkId, AcIndex aci, const 
         return false;
     }
 
-    // DoGetDelayUntilAccessRequest has already checked if the main PHY is receiving an ICF
+    // DoGetDelayUntilAccessRequest has already checked if the main PHY is receiving an ICF and
+    // above it is checked whether we can request another switch while already switching
     if (const auto state = mainPhy->GetState()->GetState();
-        state != WifiPhyState::IDLE && state != WifiPhyState::CCA_BUSY && state != WifiPhyState::RX)
+        state != WifiPhyState::IDLE && state != WifiPhyState::CCA_BUSY &&
+        state != WifiPhyState::RX && state != WifiPhyState::SWITCHING)
     {
         NS_LOG_DEBUG("Cannot request main PHY to switch when in state " << state);
         return false;
     }
 
-    // request to switch main PHY if we expect the main PHY to get channel access on this link more
-    // quickly, i.e., if ALL the ACs with queued frames (that can be transmitted on the link on
-    // which the main PHY is currently operating) and with priority higher than or equal to that of
-    // the AC for which Aux PHY gained TXOP have their backoff counter greater than the maximum
-    // between the expected delay in gaining channel access and the channel switch delay (plus PIFS
-    // if we cannot use aux PHY CCA)
+    // if the AC that is about to get channel access on the aux PHY link has no frames to send on
+    // that link, do not request the main PHY to switch
+    if (!GetStaMac()->GetQosTxop(aci)->HasFramesToTransmit(linkId))
+    {
+        NS_LOG_DEBUG("No frames of " << aci << " to send on link " << +linkId);
+        return false;
+    }
 
-    auto requestSwitch = false;
+    // if user has configured to skip the check related to the expected channel access time on
+    // the main PHY link and the AC that is about to gain access on the aux PHY link has a priority
+    // greater than or equal to the minimum priority that has been configured, switch the main PHY
+    if (!m_checkAccessOnMainPhyLink && (aci >= m_minAcToSkipCheckAccess))
+    {
+        NS_LOG_DEBUG("Skipping check related to the expected channel access time on main PHY link");
+        return true;
+    }
+
+    const auto mainPhyLinkId = GetStaMac()->GetLinkForPhy(mainPhy);
+    if (!mainPhyLinkId.has_value())
+    {
+        NS_ASSERT(m_mainPhySwitchInfo.disconnected);
+        NS_LOG_DEBUG("The main PHY is not connected to any link");
+        // we don't know when the main PHY will be connected to the link it is switching to, nor
+        // which backoff value it will possibly generate; therefore, request it to switch to the
+        // aux PHY link
+        return true;
+    }
+
+    // let AC X be the AC that is about to gain channel access on the aux PHY link, request to
+    // switch the main PHY if we do not expect any AC, with priority higher than or equal to that
+    // of AC X and with frames to send on the main PHY link, to gain channel access on the main PHY
+    // link before AC X is able to start transmitting on the aux PHY link.
+
     const auto now = Simulator::Now();
 
     for (const auto& [acIndex, ac] : wifiAcList)
     {
-        if (auto edca = GetStaMac()->GetQosTxop(acIndex);
-            acIndex >= aci && edca->HasFramesToTransmit(linkId))
+        // ignore ACs with lower priority than the AC that is about to get access on aux PHY link
+        if (acIndex < aci)
         {
-            requestSwitch = true;
+            continue;
+        }
 
-            const auto backoffEnd =
-                GetStaMac()->GetChannelAccessManager(*mainPhyLinkId)->GetBackoffEndFor(edca);
-            NS_LOG_DEBUG("Backoff end for " << acIndex
-                                            << " on preferred link: " << backoffEnd.As(Time::US));
+        const auto edca = GetStaMac()->GetQosTxop(acIndex);
+        const auto backoffEnd =
+            GetStaMac()->GetChannelAccessManager(*mainPhyLinkId)->GetBackoffEndFor(edca);
+        NS_LOG_DEBUG("Backoff end for " << acIndex
+                                        << " on main PHY link: " << backoffEnd.As(Time::US));
 
-            if ((backoffEnd <= now + minDelay) && edca->HasFramesToTransmit(*mainPhyLinkId))
-            {
-                requestSwitch = false;
-                break;
-            }
+        if ((backoffEnd <= now + minDelay) && edca->HasFramesToTransmit(*mainPhyLinkId))
+        {
+            NS_LOG_DEBUG(acIndex << " is expected to gain access on link " << +mainPhyLinkId.value()
+                                 << " sooner than " << aci << " on link " << +linkId);
+            return false;
         }
     }
 
-    return requestSwitch;
+    return true;
 }
 
 void
@@ -759,15 +922,21 @@ AdvancedEmlsrManager::SwitchMainPhyIfTxopGainedByAuxPhy(uint8_t linkId, AcIndex 
                       DONT_REQUEST_ACCESS,
                       EmlsrUlTxopAuxPhyNotTxCapableTrace(aci, Time{0}, remNav));
 
-        // the main PHY must stay for some time on this link to check if it gets channel access.
-        // The timer is stopped if a DL or UL TXOP is started. When the timer expires, the main PHY
-        // switches back to the preferred link if SwitchAuxPhy is false
-        m_switchMainPhyBackEvent.Cancel();
-        m_switchMainPhyBackEvent =
-            Simulator::Schedule(mainPhy->GetChannelSwitchDelay() + m_switchMainPhyBackDelay,
-                                &AdvancedEmlsrManager::SwitchMainPhyBackDelayExpired,
-                                this,
-                                linkId);
+        // if SwitchAuxPhy is false, the main PHY must stay for some time on this link to check if
+        // it gets channel access. The timer is stopped if a DL or UL TXOP is started. When the
+        // timer expires, the main PHY switches back to the preferred link
+        if (!m_switchAuxPhy)
+        {
+            m_switchMainPhyBackEvent.Cancel();
+            m_switchMainPhyBackEvent =
+                Simulator::Schedule(mainPhy->GetChannelSwitchDelay() + m_switchMainPhyBackDelay,
+                                    &AdvancedEmlsrManager::SwitchMainPhyBackDelayExpired,
+                                    this,
+                                    linkId,
+                                    std::nullopt);
+            // start checking PHY activity on the link the main PHY is switching to
+            RegisterListener(auxPhy);
+        }
         return;
     }
 
@@ -788,32 +957,33 @@ AdvancedEmlsrManager::SwitchMainPhyIfTxopGainedByAuxPhy(uint8_t linkId, AcIndex 
     // case, do nothing because the main PHY will start a TXOP and at the end of such TXOP
     // links will be unblocked and the channel access requested on all links
 
-    Time delay{};
+    std::optional<Time> delay;
 
     if (m_ccaLastPifs.IsPending() || m_switchMainPhyBackEvent.IsPending())
     {
         delay = std::max(Simulator::GetDelayLeft(m_ccaLastPifs),
                          Simulator::GetDelayLeft(m_switchMainPhyBackEvent));
     }
-    else if (mainPhy->IsStateSwitching() || mainPhy->IsStateCcaBusy() || mainPhy->IsStateRx())
+    else if (mainPhy->GetState()->GetLastTime(
+                 {WifiPhyState::SWITCHING, WifiPhyState::CCA_BUSY, WifiPhyState::RX}) ==
+             Simulator::Now())
     {
         delay = mainPhy->GetDelayUntilIdle();
-        NS_ASSERT(delay.IsStrictlyPositive());
     }
 
     NS_LOG_DEBUG("Main PHY state is " << mainPhy->GetState()->GetState());
     auto edca = GetStaMac()->GetQosTxop(aci);
     edca->NotifyChannelReleased(linkId); // to set access to NOT_REQUESTED
 
-    if (delay.IsZero())
+    if (!delay.has_value())
     {
         NS_LOG_DEBUG("Do nothing");
         return;
     }
 
     NS_LOG_DEBUG("Schedule channel access request on link "
-                 << +linkId << " at time " << (Simulator::Now() + delay).As(Time::NS));
-    Simulator::Schedule(delay, [=]() {
+                 << +linkId << " at time " << (Simulator::Now() + *delay).As(Time::NS));
+    Simulator::Schedule(*delay, [=]() {
         edca->StartAccessAfterEvent(linkId,
                                     Txop::DIDNT_HAVE_FRAMES_TO_TRANSMIT,
                                     Txop::CHECK_MEDIUM_BUSY);
@@ -882,52 +1052,16 @@ AdvancedEmlsrManager::SwitchMainPhyIfTxopToBeGainedByAuxPhy(uint8_t linkId,
     // use aux PHY CCA (if allowed) if the backoff has already counted down to zero on the aux PHY
     // link when the main PHY completes the switch
     const auto edca = GetStaMac()->GetQosTxop(aci);
-    const auto cam = GetStaMac()->GetChannelAccessManager(linkId);
     const auto auxPhy = GetStaMac()->GetWifiPhy(linkId);
     const auto switchDelay = mainPhy->GetChannelSwitchDelay();
-    const auto now = Simulator::Now();
     const auto auxPhyCcaCanBeUsed =
         m_useAuxPhyCca || (GetChannelForAuxPhy(linkId).GetTotalWidth() >=
                            GetChannelForMainPhy(linkId).GetTotalWidth());
-    const auto backoffEndBeforeSwitch = (cam->GetBackoffEndFor(edca) - now) <= switchDelay;
-
-    if (auxPhyCcaCanBeUsed && backoffEndBeforeSwitch)
-    {
-        NS_LOG_DEBUG("Schedule CCA check at the end of main PHY switch");
-        m_ccaLastPifs = Simulator::Schedule(switchDelay, [=, this]() {
-            // check NAV and CCA only if the backoff actually counted down to zero already
-            if (cam->GetBackoffEndFor(edca) <= Simulator::Now())
-            {
-                CheckNavAndCcaLastPifs(auxPhy, linkId, edca);
-            }
-        });
-    }
-
-    Time remNav{0};
-    if (const auto mainPhyLinkId = GetStaMac()->GetLinkForPhy(mainPhy))
-    {
-        auto mainPhyNavEnd = GetStaMac()->GetChannelAccessManager(*mainPhyLinkId)->GetNavEnd();
-        remNav = Max(remNav, mainPhyNavEnd - now);
-    }
-
-    SwitchMainPhy(linkId,
-                  false,
-                  DONT_REQUEST_ACCESS,
-                  EmlsrUlTxopAuxPhyNotTxCapableTrace(aci, delay, remNav));
-
-    // the main PHY must stay for some time on this link to check if it gets channel access. The
-    // timer is stopped if a DL or UL TXOP is started. When the timer expires, the main PHY switches
-    // back to the preferred link if SwitchAuxPhy is false
-    m_switchMainPhyBackEvent.Cancel();
-    m_switchMainPhyBackEvent =
-        Simulator::Schedule(switchDelay + m_switchMainPhyBackDelay,
-                            &AdvancedEmlsrManager::SwitchMainPhyBackDelayExpired,
-                            this,
-                            linkId);
 
     // check expected channel access delay when switch is completed
     Simulator::Schedule(switchDelay, [=, this]() {
-        const auto accessDelay = cam->GetBackoffEndFor(edca) - Simulator::Now();
+        // this is scheduled before starting the main PHY switch, hence it is executed before the
+        // main PHY is connected to the aux PHY link
 
         if (!m_switchAuxPhy && !m_switchMainPhyBackEvent.IsPending())
         {
@@ -936,19 +1070,41 @@ AdvancedEmlsrManager::SwitchMainPhyIfTxopToBeGainedByAuxPhy(uint8_t linkId,
             return;
         }
 
-        if (auxPhyCcaCanBeUsed && backoffEndBeforeSwitch && accessDelay.IsNegative())
-        {
-            // backoff already counted down to zero and we used aux PHY CCA
-            return;
-        }
-
+        const auto backoffEnd =
+            GetStaMac()->GetChannelAccessManager(linkId)->GetBackoffEndFor(edca);
         const auto pifs = GetStaMac()->GetWifiPhy(linkId)->GetPifs();
+        const auto now = Simulator::Now();
 
-        // if the remaining backoff time is shorter than PIFS when the main PHY completes the
-        // switch, we need to schedule a CCA check a PIFS after the end of the main PHY switch
-        if (accessDelay <= pifs)
+        // In case aux PHY CCA can be used and the backoff has not yet reached zero, no NAV and CCA
+        // check is needed. The channel width that will be used is the width of the aux PHY if less
+        // than a PIFS remains until the backoff reaches zero, and the width of the main PHY,
+        // otherwise. If aux PHY CCA can be used and the backoff has already reached zero, a NAV and
+        // CCA check is needed.
+
+        if (auxPhyCcaCanBeUsed && backoffEnd < now)
         {
-            // use main PHY CCA in the last PIFS interval after main PHY switch end
+            /**
+             * use aux PHY CCA in the last PIFS interval before main PHY switch end
+             *
+             *        Backoff    Switch
+             *          end     end (now)
+             * ──────────┴─────────┴──────────
+             *      |---- PIFS ----|
+             */
+            CheckNavAndCcaLastPifs(auxPhy, linkId, edca);
+        }
+        else if (!auxPhyCcaCanBeUsed && (backoffEnd - now <= pifs))
+        {
+            /**
+             * the remaining backoff time (if any) when the main PHY completes the switch is shorter
+             * than or equal to a PIFS, thus the main PHY performs CCA in the last PIFS interval
+             * after switch end
+             *
+             *        Switch    Backoff                 Backoff    Switch
+             *       end (now)    end                     end     end (now)
+             * ──────────┴─────────┴──────────   ──────────┴─────────┴──────────
+             *           |---- PIFS ----|                            |---- PIFS ----|
+             */
             NS_LOG_DEBUG("Schedule CCA check a PIFS after the end of main PHY switch");
             m_ccaLastPifs = Simulator::Schedule(pifs,
                                                 &AdvancedEmlsrManager::CheckNavAndCcaLastPifs,
@@ -957,15 +1113,45 @@ AdvancedEmlsrManager::SwitchMainPhyIfTxopToBeGainedByAuxPhy(uint8_t linkId,
                                                 linkId,
                                                 edca);
         }
-        else if (!m_switchAuxPhy &&
-                 !GetExpectedAccessWithinDelay(linkId,
-                                               Simulator::GetDelayLeft(m_switchMainPhyBackEvent) +
-                                                   mainPhy->GetChannelSwitchDelay()))
+        else if (WifiExpectedAccessReason reason;
+                 !m_switchAuxPhy &&
+                 (reason = GetStaMac()->GetChannelAccessManager(linkId)->GetExpectedAccessWithin(
+                      Simulator::GetDelayLeft(m_switchMainPhyBackEvent) +
+                      mainPhy->GetChannelSwitchDelay())) !=
+                     WifiExpectedAccessReason::ACCESS_EXPECTED)
         {
             NS_LOG_DEBUG("No AC is expected to get backoff soon, switch main PHY back");
-            SwitchMainPhyBackDelayExpired(linkId);
+            SwitchMainPhyBackDelayExpired(linkId, reason);
         }
     });
+
+    Time remNav{0};
+    if (const auto mainPhyLinkId = GetStaMac()->GetLinkForPhy(mainPhy))
+    {
+        auto mainPhyNavEnd = GetStaMac()->GetChannelAccessManager(*mainPhyLinkId)->GetNavEnd();
+        remNav = Max(remNav, mainPhyNavEnd - Simulator::Now());
+    }
+
+    SwitchMainPhy(linkId,
+                  false,
+                  DONT_REQUEST_ACCESS,
+                  EmlsrUlTxopAuxPhyNotTxCapableTrace(aci, delay, remNav));
+
+    // if SwitchAuxPhy is false, the main PHY must stay for some time on this link to check if it
+    // gets channel access. The timer is stopped if a DL or UL TXOP is started. When the timer
+    // expires, the main PHY switches back to the preferred link
+    if (!m_switchAuxPhy)
+    {
+        m_switchMainPhyBackEvent.Cancel();
+        m_switchMainPhyBackEvent =
+            Simulator::Schedule(switchDelay + m_switchMainPhyBackDelay,
+                                &AdvancedEmlsrManager::SwitchMainPhyBackDelayExpired,
+                                this,
+                                linkId,
+                                std::nullopt);
+        // start checking PHY activity on the link the main PHY is switching to
+        RegisterListener(auxPhy);
+    }
 }
 
 } // namespace ns3